Add files via upload
How to use?

##########################################################

Hello, Run the application. First scan the QR Code. Then scan the desired inscription. (Be patient because the scanning process must read and query the article.) The software will continue processing according to the QR Code you have read. If you scan the QR Code other than TEAS|M99999|H99999|I99999 or TEAS|M99999|H99999|I99999|ITAAAA, you may receive an error. AAAA and 99999 are representative. After the successfully transferred article, the program will ask you if you will scan it again. Press ENTER to continue scanning. If you want to send the scanned files as Mail, you can start again the IETApp. Please don't forget to share your feedback with us. Finally, it is forbidden to make unauthorized changes to the codes of this software, make a copy of it and put it up for sale by changing the name, or use it for commercial purposes.

Explain for The IETApp

##########################################################

Hello users and my fellow software developers, Reading the Job Tracking Form is a routine work we do at the workplace every day, getting the necessary information and saving this information in the appropriate place in the Excel Workspace. Then, to date, these recorded files and send us an e-mail to the required person. We only ask the user to show the QR Code and Required Information on the Job Tracking Form to the camera. In this way, users have these magic codes to do the long processes I mentioned above. If you follow the instructions that appear on the screen correctly while using the program, the program can't fail. You can use it with peace of mind.

Updates

##########################################################

Version 1.0.2 Erase update of scanned data outputs after all operations so that the data is not mixed. Tue 12 Apr 09:25 P.M

Version 1.0.3 Add sound effects update. Wed 13 Apr 08:50 A.M.

Version 1.0.4 Performance Updates Thu 14 Apr 05.06 P.M.

Version 1.0.5 Text Reading Performance Update Sun 17 Apr 11.20 A.M.

##########################################################

| D e v e l o p e d by C a f t e l l e
| Caftelle created by Furkan ARINCI | Copyright 2022